api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.0.1
python-multipart>=0.0.9
pydantic==2.5.0
gradio==4.44.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
